Committing changes lost in a backup restore on the server.
2008-09-01: The check for advanced/normal install should only be carried out if the page to select the type of install is included. This is now the case.
I've converted the evar pages to use nsDialogs instead of InstallOptions. The reason for this was that I wanted to be able to make the "Browse for file" dialog a little smarter about what directory it started looking in.

2008-09-03: Updated the language strings. Mostly the French ones.

2008-09-15: Updated to xPaintClass 1.0 update 5.
Made it possible to compile without support for registering extensions while keeping the uninstaller.
Fixed a bug in PageAndSectionConfigEvars.nsh where it wouldn't save changes to evars that were made manually (as apposed to using the Browse button).

2008-09-16: Added SLI-ModuleManager and settingsipc-1.0.0.1.dll (needed by sli-modulemanager).
Released LOSI 0.3